move setup and update wsgi reference
