+++ setup/apache-setup/002-tworavens.conf
@@ -1,0 +1,37 @@
+<VirtualHost *:8080>
+  ServerName two_ravens_docker
+
+  ServerAdmin webmaster@localhost
+
+  Alias /static /var/www/html/2ravens/static/
+  <Directory /var/www/html/2ravens/static>
+    Require all granted
+  </Directory>
+
+  <Location /RApacheInfo>
+    SetHandler r-info
+  </Location>
+
+  WSGIDaemonProcess tworavens_django processes=2 threads=12 python-home=/root/virtualenvs/2ravens python-path=/srv/webapps/TwoRavens
+  WSGIProcessGroup tworavens_django
+  WSGIApplicationGroup %{GLOBAL}
+  WSGIScriptAlias / /srv/webapps/TwoRavens/tworavensproject/wsgi_dev_container.py
+
+  <Directory /srv/webapps/TwoRavens/tworavensproject>
+      <Files wsgi_dev_container.py>
+          Require all granted
+      </Files>
+  </Directory>
+
+  ErrorLog ${APACHE_LOG_DIR}/error.log
+  CustomLog ${APACHE_LOG_DIR}/access.log combined
+
+  #DocumentRoot /var/www/html
+
+  # Available loglevels: trace8, ..., trace1, debug, info, notice, warn,
+  # error, crit, alert, emerg.
+  # It is also possible to configure the loglevel for particular
+  # modules, e.g.
+  #LogLevel info ssl:warn
+
+</VirtualHost>
