+++ setup/apache-setup/README.md
@@ -1,0 +1,7 @@
+Reference material for apache setup with Django and RApache
+
+
+- File: `002-tworavens.conf`
+    - Initial apache test file for use on a docker container running ubuntu 16.04
+
+-
